Fix issue with -P option and text window restoration.
When the -P option was enabled, the VDU command to set the text window
was being written to the file towards the start, causing the whole
screen contents to be displayed in the window.  This switches this to
the end, i.e. the text window is written to the file at the same time it
is being restored after the screen contents have been written.