Fix extra blank lines in -P mode.
In -P mode the intention is that feeding the file to the VDU drivers,
for example with *PRINT, will restore the screen to how it was.  With
lines that occupy the full width of the screen this was not the case as
printing in the last column caused a wrap to the next line and then the
newline (CR,LF) caused a move to the following line, leaving a blank
line in between.  This commit fixes this by omitting the newline in -P
mode if the last character position on a line is occupied.